Restructured folder and created add to cart button component for more maintainable code
