+++ src/app/types/currency.tsx
@@ -1,0 +1,1 @@
+export type Currency = "USD" | "GBP" | "EUR" | "JPY";
